спасительный коммит

+++ src/main/java/com/boiko_ivan/spring/levelup_back/repositories/UserRepository.java
@@ -1,10 +1,11 @@
 package com.boiko_ivan.spring.levelup_back.repositories;
 
+
 import com.boiko_ivan.spring.levelup_back.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
     Optional<User> findByEmail(String email);
 }
